refactor export pipeline into internal modules
