+++ .gitignore
@@ -3,3 +3,5 @@
 
 # macOS
 .DS_Store
+
+/scripts
