+++ Makefile
@@ -2,7 +2,7 @@ CC = gcc
 CFLAGS = -Wall -Wextra -Wshadow -Wcast-align -Wwrite-strings -Wredundant-decls \
          -Wstrict-prototypes -Wold-style-definition -std=c99 -O2 -D_POSIX_C_SOURCE=200809L
 TARGET = fuori
-SOURCES = fuori.c ignore.c
+SOURCES = main.c collect.c render.c git_paths.c ignore.c
 PREFIX ?= /usr/local
 BINDIR ?= $(PREFIX)/bin
 
@@ -18,4 +18,4 @@ install: $(TARGET)
 	install -d $(BINDIR)
 	install -m 755 $(TARGET) $(BINDIR)
 
-.PHONY: all clean install
+.PHONY: all clean test install
